api/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudinary==1.44.1
cryptography==46.0.3
distro==1.9.0
docstring_parser==0.17.0
fal_client==0.8.1
fastapi==0.121.0
ffmpeg-python==0.2.0
filelock==3.20.0
flasgger==0.9.7.1
Flask==3.0.0
flask-cors==6.0.1
flask-swagger-ui==5.21.0
fsspec==2025.9.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-genai==1.46.0
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mcp==1.19.0
mistune==3.1.4
oauthlib==3.3.1
openai==2.6.0
packaging==25.0
pillow==12.0.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyMySQL==1.1.2
pyparsing==3.2.5
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.28.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sse-starlette==3.0.2
starlette==0.49.1
tenacity==9.1.2
tokenizers==0.22.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
